% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phac_style.R
\name{phac_style}
\alias{phac_style}
\title{Apply the PHAC style to a ggplot object.}
\usage{
phac_style(
  g,
  size.base = 12,
  palette = "Safe",
  color.type = "d",
  font.family = "sans"
)
}
\arguments{
\item{g}{ggplot object.}

\item{size.base}{Integer. Base size for text in plot. Default: \code{size.base=12}}

\item{palette}{String. Color palette name used in the \code{rcartocolor} package.}

\item{color.type}{String. \code{'d'} for discrete (default), \code{'c'} for continuous}

\item{font.family}{String. Name of family fonts used. Default: \code{'sans'}.}
}
\value{
A ggplot object with the update style.
}
\description{
Apply the PHAC style to a ggplot object.
}
\examples{

# Create the base plot
data(Orange)
g = ggplot(data = Orange,
           aes(x=age, y=circumference, color=factor(Tree))) +
  geom_line(size=1) +
  labs(
    title   = 'Orange growth',
    caption = 'Data source: from the `datasets` package',
    color   = 'Tree ID'
  )

# Standard format from ggplot2:
plot(g)

# Improved style:
plot(phac_style(g))
plot(phac_style(g + facet_wrap(~Tree)))


}
